requirements.txt
certifi==2022.5.18.1
Flask==2.0.3
Flask_RESTful==0.3.9
gevent_socketio==0.3.6
gunicorn==20.1.0
pyjwt==2.4.0
pymongo==3.12.0
pymongo[srv]
python-dotenv==0.19.2
python_dateutil==2.8.2
pytz==2022.1
requests==2.26.0
setuptools==61.2.0
six==1.16.0
urllib3==1.26.9
git+https://github.com/paramatrixtech/ksapi.git#egg=ks_api_client&subdirectory=./python
